refactor(*): 重构了大部分代码 添加了几个用户服务，发送邮箱验证码服务 与之前很多功能都不一样了

+++ rare-earth-track/rare-earth-track-mgb/src/main/java/com/rare_earth_track/mgb/mapper/RetResourceRoleRelationMapper.java
@@ -25,7 +25,7 @@ public interface RetResourceRoleRelationMapper {
      * @param id id
      * @return int
      */
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     /**
      * insert
@@ -53,7 +53,7 @@ public interface RetResourceRoleRelationMapper {
      * @param id id
      * @return com.rare_earth_track.mgb.model.RetResourceRoleRelation
      */
-    RetResourceRoleRelation selectByPrimaryKey(Integer id);
+    RetResourceRoleRelation selectByPrimaryKey(Long id);
 
     /**
      * updateByExampleSelective
